Enable many more warnings, and fix them.

+++ core.h
@@ -54,7 +54,7 @@ void doc_end(struct DocBuilder *builder);
 void doc_bracket_open(struct DocBuilder *builder, const char *left);
 void doc_bracket_close(struct DocBuilder *builder, const char *right);
 
-void pretty(FILE *file, int width, struct Doc *docs, int length);
+void pretty(FILE *file, size_t width, struct Doc *docs, int length);
 void dump_docs(struct Doc *docs, int length);
 
 #endif
